requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
scikit-learn==1.4.0
numpy==1.26.3
gunicorn==21.2.0
python-dotenv==1.0.0
